Fix: Improve dark mode text contrast in Interactive Tutorials tab
- Changed heading color from dark:text-black to dark:text-white for visibility in dark mode
- Updated description color from dark:text-gray-600 to dark:text-gray-300 for better readability
- Ensured all text in the Interactive Tutorials section is accessible in both light and dark themes
